// src/types.ts
export type championData = {
  displayName: string,
  id: string,
};
// ...
export interface PlayerStats {
  puuid: string;
  gameName: string;
  tagLine: string;
  profileIconId: number;
  summonerLevel: number; 
  matchesPlayed: number;
  latestGamePlayed: EpochTimeStamp;
  placements: placementDto
  placementAvg: number;
  infographics: infographicsDto;
  augmentStats: augmentsStatsDto;
  championStats: {
    [championName: string]: championStatsDto;
  }
};

export interface placementDto {
  [key: number]: number;
};

export interface championStatsDto {
  timesPlayed: number;
  placements: placementDto;
  placementAvg: number;
  infographics: infographicsDto;
  augmentStats: augmentsStatsDto;
  name: string;
  id: string;
  stage: number;
};

export interface infographicsDto {
  damageStats: damageStatsDto;
  damageTakenStats: damageTakenStatsDto;
  goldStats: goldStatsDto;
  skillShotsStats: skillShotsDto;
  killsDeathsAssists: killDeathAssistsDto;
};

export interface augmentsStatsDto {
  [id: number]: {
    picked: number;
  }
}

export interface killDeathAssistsDto {
  kda: numericalStatsDto;
  kills: numericalStatsDto;
  deaths: numericalStatsDto;
  assists: numericalStatsDto;
}

export interface damageStatsDto {
  total: {
    total: numericalStatsDto;
    champions: numericalStatsDto;
  };
  true: {
    total: numericalStatsDto;
    champions: numericalStatsDto;
  };
  magic: {
    total: numericalStatsDto;
    champions: numericalStatsDto;
  };
  physical: {
    total: numericalStatsDto;
    champions: numericalStatsDto;
  };
  perMinute: numericalStatsDto;
};

export interface damageTakenStatsDto {
  total: numericalStatsDto;
  true: numericalStatsDto;
  magic: numericalStatsDto;
  physical: numericalStatsDto;
  mitigated: numericalStatsDto;
};

export interface skillShotsDto {
  dodged: numericalStatsDto;
  hit: numericalStatsDto;
};

export interface goldStatsDto {
  earned: numericalStatsDto;
  spent: numericalStatsDto;
  perMinute: numericalStatsDto;
};

export interface numericalStatsDto {
  value: number;
  records: recordsDto;
};

export interface recordsDtoEntry {
  value: number;
  matchId: string;
}

export type recordsDto = recordsDtoEntry[];